api/requirements.txt
aiohttp==3.8.1
python-dotenv==0.20.0
asyncpg==0.25.0
zope.sqlalchemy==1.6